[MOD/#378] 패키징 수정
