+++ core/network/src/main/java/com/hilingual/core/network/model/BaseResponse.kt
@@ -1,0 +1,14 @@
+package com.hilingual.core.network.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BaseResponse<T>(
+    @SerialName("code")
+    val code: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: T? = null
+)
