+++ data/auth/src/main/java/com/hilingual/data/auth/service/ReissueService.kt
@@ -15,8 +15,8 @@
  */
 package com.hilingual.data.auth.service
 
-import com.hilingual.core.network.AUTHORIZATION
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.constant.AUTHORIZATION
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.auth.dto.response.ReissueTokenResponseDto
 import retrofit2.http.Header
 import retrofit2.http.POST
