+++ data/auth/src/main/java/com/hilingual/data/auth/service/VerifyService.kt
@@ -1,6 +1,6 @@
 package com.hilingual.data.auth.service
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.auth.dto.request.VerifyCodeRequestDto
 import retrofit2.http.Body
 import retrofit2.http.POST
