+++ data/calendar/src/main/java/com/hilingual/data/calendar/datasouceimpl/CalendarRemoteDataSourceImpl.kt
@@ -15,7 +15,7 @@
  */
 package com.hilingual.data.calendar.datasouceimpl
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.calendar.datasource.CalendarRemoteDataSource
 import com.hilingual.data.calendar.dto.response.CalendarResponseDto
 import com.hilingual.data.calendar.dto.response.DiaryThumbnailResponseDto
