+++ data/calendar/src/main/java/com/hilingual/data/calendar/service/CalendarService.kt
@@ -15,7 +15,7 @@
  */
 package com.hilingual.data.calendar.service
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.calendar.dto.response.CalendarResponseDto
 import com.hilingual.data.calendar.dto.response.DiaryThumbnailResponseDto
 import com.hilingual.data.calendar.dto.response.TopicResponseDto
