+++ data/diary/src/main/java/com/hilingual/data/diary/service/DiaryService.kt
@@ -15,7 +15,7 @@
  */
 package com.hilingual.data.diary.service
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.diary.dto.request.BookmarkRequestDto
 import com.hilingual.data.diary.dto.request.DiaryFeedbackCreateRequestDto
 import com.hilingual.data.diary.dto.response.DiaryContentResponseDto
