+++ data/feed/src/main/java/com/hilingual/data/feed/service/FeedService.kt
@@ -1,6 +1,6 @@
 package com.hilingual.data.feed.service
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.feed.dto.request.LikeRequestDto
 import com.hilingual.data.feed.dto.response.DiaryProfileResponseDto
 import com.hilingual.data.feed.dto.response.FeedProfileResponseDto
