+++ data/presigned/src/main/java/com/hilingual/data/presigned/di/ServiceModule.kt
@@ -1,6 +1,6 @@
 package com.hilingual.data.presigned.di
 
-import com.hilingual.core.network.NoAuthClient
+import com.hilingual.core.network.di.NoAuthClient
 import com.hilingual.data.presigned.service.PresignedUrlService
 import com.hilingual.data.presigned.service.S3Service
 import dagger.Module
