+++ data/user/src/main/java/com/hilingual/data/user/datasource/UserRemoteDataSource.kt
@@ -15,7 +15,7 @@
  */
 package com.hilingual.data.user.datasource
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.user.dto.response.follow.FollowerResponseDto
 import com.hilingual.data.user.dto.response.follow.FollowingResponseDto
 import com.hilingual.data.user.dto.response.notification.NotificationDetailResponseDto
