+++ data/voca/src/main/java/com/hilingual/data/voca/datasource/VocaDataSource.kt
@@ -15,7 +15,7 @@
  */
 package com.hilingual.data.voca.datasource
 
-import com.hilingual.core.network.BaseResponse
+import com.hilingual.core.network.model.BaseResponse
 import com.hilingual.data.voca.dto.response.VocaDetailResponseDto
 import com.hilingual.data.voca.dto.response.VocaListResponseDto
